+++ lib/constants.js
@@ -1,0 +1,9 @@
+const xParserMessageName = 'x-parser-message-name';
+const xParserSchemaId = 'x-parser-schema-id';
+const xParserCircle = 'x-parser-circular';
+const xParserCircleProps = 'x-parser-circular-props';
+
+module.exports.xParserMessageName = xParserMessageName;
+module.exports.xParserSchemaId = xParserSchemaId;
+module.exports.xParserCircle = xParserCircle;
+module.exports.xParserCircleProps = xParserCircleProps;
